Specify the board as GENERIC_RP_RP2040
The default board seems to be GENERIC_PROMICRO_RP2040, which defines
the RP2040_BOOTLOADER_DOUBLE_TAP_RESET constant, but GENERIC_RP_RP2040
does not.

This was causing issues when switching between computers with a usb
switch. It's a 4 position switch, with 2 positions "filled", and a single button
to cycle through the 4 positions. So to switch computers, I would press the
button twice in quick succession, but this was triggering the
DOUBLE_TAP_RESET functionality.